% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycan_analysis.R
\name{glycan_analysis_structures}
\alias{glycan_analysis_structures}
\title{Get GlycanAnalysis structures}
\usage{
glycan_analysis_structures(analysis)
}
\arguments{
\item{analysis}{GlycanAnalysis object}
}
\value{
List of predicted structures
}
\description{
Get GlycanAnalysis structures
}
