% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycan.R
\name{glycan_mass}
\alias{glycan_mass}
\title{Calculate glycan mass}
\usage{
glycan_mass(glycan_code)
}
\arguments{
\item{glycan_code}{4-5 digit composition code}
}
\value{
Numeric mass in Daltons
}
\description{
Calculate mass from a glycan composition code.
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
mass <- glycan_mass("4501")
}
}
