% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycan.R
\name{glycan_new}
\alias{glycan_new}
\title{Create a new Glycan object}
\usage{
glycan_new(glycan_code, glycan_type = "N")
}
\arguments{
\item{glycan_code}{4-5 digit composition code (e.g., "4501" for HexNAc-4, Hex-5, Fuc-0, NeuAc-1)}

\item{glycan_type}{Type of glycan: "N" for N-glycan or "O" for O-glycan (default: "N")}
}
\value{
Python Glycan object
}
\description{
Create a new Glycan object
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
glycan <- glycan_new("4501", "N")
}
}
