% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycan.R
\name{glycan_predict_structures}
\alias{glycan_predict_structures}
\title{Predict glycan structures}
\usage{
glycan_predict_structures(glycan)
}
\arguments{
\item{glycan}{Glycan object created by glycan_new()}
}
\value{
List of predicted structure graphs
}
\description{
Predict possible structures from a glycan composition.
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
glycan <- glycan_new("4501", "N")
structures <- glycan_predict_structures(glycan)
}
}
