% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{glycofragR}
\alias{glycofragR-package}
\alias{glycofragR}
\title{glycofragR}
\description{
R wrapper for the glycofrag Python package.
}
\author{
\strong{Maintainer}: Oluwatosin Daramola \email{oluwatosindaramola@gmail.com}

}
