% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycopeptide.R
\name{glycopeptide_fragments}
\alias{glycopeptide_fragments}
\title{Generate glycopeptide fragments}
\usage{
glycopeptide_fragments(glycopeptide)
}
\arguments{
\item{glycopeptide}{Glycopeptide object created by glycopeptide_new()}
}
\value{
List of fragment ions (peptide, glycan, and glycopeptide fragments)
}
\description{
Generate all theoretical fragment ions for a glycopeptide.
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
gp <- glycopeptide_new("LCPDCPLLAPLNDSR", "4501", 12, "N")
frags <- glycopeptide_fragments(gp)
}
}
