% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycopeptide.R
\name{glycopeptide_new}
\alias{glycopeptide_new}
\title{Create a new Glycopeptide object}
\usage{
glycopeptide_new(
  peptide_sequence,
  glycan_code,
  glycosylation_site,
  glycan_type = "N",
  modifications = NULL,
  mod_string = NULL
)
}
\arguments{
\item{peptide_sequence}{Amino acid sequence}

\item{glycan_code}{4-5 digit glycan composition code}

\item{glycosylation_site}{Position of glycosylation (1-based indexing)}

\item{glycan_type}{Type of glycan: "N" or "O" (default: "N")}

\item{modifications}{Optional named list of peptide modifications with positions (1-based)}

\item{mod_string}{Optional modification string (e.g., "M:Ox; C:CAM")}
}
\value{
Python Glycopeptide object
}
\description{
Create a new Glycopeptide object
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
gp <- glycopeptide_new(
  peptide_sequence = "LCPDCPLLAPLNDSR",
  glycan_code = "4501",
  glycosylation_site = 12,
  glycan_type = "N"
)
}
}
