% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycopeptide.R
\name{glycopeptide_summary}
\alias{glycopeptide_summary}
\title{Get glycopeptide summary}
\usage{
glycopeptide_summary(glycopeptide)
}
\arguments{
\item{glycopeptide}{Glycopeptide object}
}
\value{
List with peptide_sequence, glycan_code, glycosylation_site, glycan_type
}
\description{
Extract key information from a glycopeptide object.
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
gp <- glycopeptide_new("LCPDCPLLAPLNDSR", "4501", 12, "N")
summary <- glycopeptide_summary(gp)
}
}
