% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_supported_modifications}
\alias{list_supported_modifications}
\title{List supported modifications}
\usage{
list_supported_modifications(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical; if TRUE, print detailed information (default: TRUE)}
}
\value{
List of supported modifications with masses and targets
}
\description{
Get a list of all supported peptide modifications.
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
mods <- list_supported_modifications()
}
}
