% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{peptide_fragments}
\alias{peptide_fragments}
\title{Generate peptide fragments}
\usage{
peptide_fragments(peptide, charge_states = c(1, 2))
}
\arguments{
\item{peptide}{Peptide object created by peptide_new()}

\item{charge_states}{Vector of charge states to generate (default: c(1, 2))}
}
\value{
List of fragment ions
}
\description{
Generate theoretical peptide fragment ions (b, y, c, z).
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
peptide <- peptide_new("PEPTIDE")
frags <- peptide_fragments(peptide, charge_states = c(1, 2, 3))
}
}
