% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_glycofrag_conda_env}
\alias{set_glycofrag_conda_env}
\title{Set glycofrag conda environment}
\usage{
set_glycofrag_conda_env(env_name = "glycofrag")
}
\arguments{
\item{env_name}{Name of the conda environment (default: "glycofrag")}
}
\value{
TRUE invisibly
}
\description{
Configure reticulate to use a specific conda environment containing glycofrag.
}
\examples{
\dontrun{
set_glycofrag_conda_env("glycofrag")
}
}
